+++ src/app/app.component.ts
@@ -1,27 +1,56 @@
-import { Component, OnDestroy, OnInit } from '@angular/core';
+import { Component, HostBinding, OnDestroy, OnInit } from '@angular/core';
 import { default as makeBlockie } from 'ethereum-blockies-base64';
 import { Subscription } from 'rxjs';
 import { ChannelPollingService } from './services/channel-polling.service';
 import { RaidenService } from './services/raiden.service';
 import { SharedService } from './services/shared.service';
+import { MediaObserver } from '@angular/flex-layout';
 
 @Component({
     selector: 'app-root',
     templateUrl: './app.component.html',
-    styleUrls: ['./app.component.css']
+    styleUrls: ['./app.component.scss']
 })
 export class AppComponent implements OnInit, OnDestroy {
+    @HostBinding('@.disabled')
+    public animationsDisabled = false;
+
     public title = 'Raiden';
     public raidenAddress;
 
-    pendingRequests = 0;
+    private _pendingRequests = 0;
+
+    get pendingRequests(): string {
+        return this._pendingRequests.toString();
+    }
+
     private sub: Subscription;
+    private _menuOpen: boolean;
 
     constructor(
         private sharedService: SharedService,
         private raidenService: RaidenService,
-        private channelPollingService: ChannelPollingService
-    ) {}
+        private channelPollingService: ChannelPollingService,
+        private media: MediaObserver
+    ) {
+        this._menuOpen = false;
+    }
+
+    get menuOpen(): boolean {
+        if (!this.isMobile()) {
+            return true;
+        } else {
+            return this._menuOpen;
+        }
+    }
+
+    toggleMenu() {
+        this._menuOpen = !this._menuOpen;
+    }
+
+    isMobile(): boolean {
+        return this.media.isActive('xs');
+    }
 
     ngOnInit() {
         this.raidenService.raidenAddress$.subscribe(
@@ -30,7 +59,7 @@ export class AppComponent implements OnInit, OnDestroy {
         this.sub = this.sharedService.pendingRequests.subscribe(
             pendingRequests => {
                 setTimeout(() => {
-                    this.pendingRequests = pendingRequests;
+                    this._pendingRequests = pendingRequests;
                 });
             }
         );
@@ -38,6 +67,16 @@ export class AppComponent implements OnInit, OnDestroy {
             .channels()
             .subscribe();
         this.sub.add(pollingSubscription);
+
+        this.disableAnimationsOnAndroid();
+    }
+
+    private disableAnimationsOnAndroid() {
+        const userAgent = navigator.userAgent.toLowerCase();
+        const isAndroid = userAgent.indexOf('android') > -1;
+        if (isAndroid) {
+            this.animationsDisabled = true;
+        }
     }
 
     ngOnDestroy() {
@@ -59,5 +98,9 @@ export class AppComponent implements OnInit, OnDestroy {
 
     attemptConnection() {
         this.raidenService.attemptConnection();
+    }
+
+    closeMenu() {
+        this._menuOpen = false;
     }
 }