+++ src/app/components/channel-table/channel-table.component.spec.ts
@@ -38,6 +38,10 @@ import { ChannelTableComponent } from './channel-table.component';
 import { StatusPipe } from '../../pipes/status.pipe';
 import Spy = jasmine.Spy;
 import { TestProviders } from '../../../testing/test-providers';
+import { SortFilterPageHeaderComponent } from '../page/sort-filter-page-header/sort-filter-page-header.component';
+import { ChannelActionsComponent } from '../channel-actions/channel-actions.component';
+import { PageBaseComponent } from '../page/page-base/page-base.component';
+import { PageItemComponent } from '../page/page-item/page-item.component';
 
 describe('ChannelTableComponent', () => {
     let component: ChannelTableComponent;
@@ -53,6 +57,10 @@ describe('ChannelTableComponent', () => {
                 OpenDialogComponent,
                 TokenInputComponent,
                 AddressInputComponent,
+                SortFilterPageHeaderComponent,
+                PageBaseComponent,
+                PageItemComponent,
+                ChannelActionsComponent,
                 TokenPipe,
                 EllipsisPipe,
                 KeysPipe,