+++ src/app/components/channel-table/channel-table.component.ts
@@ -6,7 +6,7 @@ import {
     trigger
 } from '@angular/animations';
 import { Component, OnDestroy, OnInit, ViewChild } from '@angular/core';
-import { MatDialog, MatPaginator, PageEvent } from '@angular/material';
+import { MatDialog, PageEvent } from '@angular/material';
 import { EMPTY, Subscription } from 'rxjs';
 import { Observable } from 'rxjs/internal/Observable';
 import { flatMap } from 'rxjs/operators';
@@ -39,6 +39,7 @@ import {
 import { ChannelSorting } from './channel.sorting.enum';
 import { AddressBookService } from '../../services/address-book.service';
 import { Addresses } from '../../models/address';
+import { PageBaseComponent } from '../page/page-base/page-base.component';
 
 @Component({
     selector: 'app-channel-table',
@@ -67,7 +68,8 @@ import { Addresses } from '../../models/address';
     ]
 })
 export class ChannelTableComponent implements OnInit, OnDestroy {
-    @ViewChild(MatPaginator) paginator: MatPaginator;
+    @ViewChild(PageBaseComponent)
+    page: PageBaseComponent;
 
     public channels$: Observable<Channel[]>;
     public amount: number;
@@ -147,13 +149,13 @@ export class ChannelTableComponent implements OnInit, OnDestroy {
 
     applyKeywordFilter() {
         this.applyFilters(this.sorting);
-        this.paginator.firstPage();
+        this.page.firstPage();
     }
 
     clearFilter() {
         this.filter = '';
         this.applyFilters(this.sorting);
-        this.paginator.firstPage();
+        this.page.firstPage();
     }
 
     // noinspection JSMethodCanBeStatic
@@ -179,9 +181,7 @@ export class ChannelTableComponent implements OnInit, OnDestroy {
         };
 
         const dialog = this.dialog.open(PaymentDialogComponent, {
-            width: '500px',
-            data: payload,
-            autoFocus: false
+            data: payload
         });
 
         dialog
@@ -209,7 +209,6 @@ export class ChannelTableComponent implements OnInit, OnDestroy {
         };
 
         const dialog = this.dialog.open(DepositDialogComponent, {
-            width: '500px',
             data: payload
         });
 
@@ -245,7 +244,6 @@ export class ChannelTableComponent implements OnInit, OnDestroy {
         };
 
         const dialog = this.dialog.open(ConfirmationDialogComponent, {
-            width: '500px',
             data: payload
         });
 
@@ -276,7 +274,6 @@ export class ChannelTableComponent implements OnInit, OnDestroy {
         };
 
         const dialog = this.dialog.open(OpenDialogComponent, {
-            width: '500px',
             data: payload
         });