+++ src/testing/mock-mat-dialog.ts
@@ -2,11 +2,14 @@ import { of } from 'rxjs';
 
 export class MockMatDialog {
     cancelled: boolean;
+    returns: () => any = () => true;
+
+    constructor() {}
 
     open() {
         return {
             afterClosed: () => {
-                return of(this.cancelled ? null : true);
+                return of(this.cancelled ? null : this.returns());
             }
         };
     }